rollback error change

+++ src/main/java/com/yali/vilivili/repository/UserRepository.java
@@ -27,7 +27,7 @@ public interface UserRepository extends JpaRepository<User, String> {
      */
     @Modifying
     @Query(value = "update user set  is_valid= ?2 where id = ?1", nativeQuery = true)
-    int deleteById(Integer id, int isValid);
+    int deleteById(Integer id, byte isValid);
 
     @Query(value = "select * from user where username = ?1 and is_valid = ?2 and type=?3",nativeQuery = true)
     List<User> findAllUser(String username, int isValid, int Type);
